Rename shape_renderer to renderer
