+++ src/core/CMakeLists.txt
@@ -6,7 +6,7 @@ find_package(glm CONFIG REQUIRED)
 
 add_library(core STATIC
     buffer.cpp
-    shape_renderer.cpp
+    renderer.cpp
     window.cpp
     shader.cpp
     texture.cpp
