+++ src/core/renderer.hpp
